feat: 봉투 생성 응답값 변경

+++ data/src/main/java/com/susu/data/remote/model/request/FriendRequest.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 data class FriendRequest(
     val name: String,
     val phoneNumber: String? = null,
-    val relationshipId: Int,
+    val relationshipId: Long,
     val customRelation: String? = null,
 )
